Add alpha diversity measures

add_alpha() adds selected alpha diversity measures to the sample table of a tidytacos object.

Usage

add_alphas(ta, methods = "all", ...)

Arguments

ta

A tidytacos object.

methods

A character vector of the diversity measure to use, see add_alpha() for examples. Optionally use 'all' to add all diversity measures.

...

Arguments passed on to add_alpha

method

The diversity measure to use, see vegan::diversity() for further information on these.

keep_empty_samples

Whether to keep empty samples or not.

Value

A tidytacos object with the selected alpha diversity measures added.

Details

This function can add multiple different alpha diversity measures to the sample table, specified by the methods argument.

See also

Examples

urt_all_alphas <- urt %>% add_alphas()
#> Warning: Removed 3 empty samples.