Adds the most dominant taxa in the sample to the sample table

add_dominant_taxa() adds the most dominant taxa in the sample above a given relative abundance to the sample table. Samples that do not have a dominant taxon will have NA in the dominant_taxon column.

Usage

add_dominant_taxa(ta, threshold_dominance = 0.5, taxon_name = taxon_id)

Arguments

ta

A tidytacos object.

threshold_dominance

The relative abundance threshold for a taxon to be considered dominant.

taxon_name

The column name of the taxa table that defines the taxon name.

Value

A tidytacos object with the dominant taxa added to the sample table and the Berger-Parker index of the most dominant taxon. If the B-P index is lower than the threshold NA is returned.

See also