Add alpha diversity measures using subsampling

add_subsampled_alpha() adds selected alpha diversity measures to the sample table of a tidytacos object using an itterative subsampling process.

Usage

add_subsampled_alpha(
  ta,
  min_lib_size = NULL,
  method = "shannon",
  itterations = 100
)

Arguments

ta

a tidytacos object.

min_lib_size

the minimum lib size samples need to have. Samples with lower lib sizes will be discarded and samples with a higher readcount will be itteratively subsampled to this readcount to allow for a fair comparison across read_depths.

method

The diversity measure to use, see vegan::diversity() for further information on these.

itterations

the amount of itterations for subsampling. Please report this number in your research.

Value

A tidytacos object with the selected alpha diversity measure added.

See also