Performs Markov Clustering on a sparcc network matrix generated by filter_network(). and returns a tibble with clusters and taxon identifiers. Optionally, the network can be visualized.

Performs Markov Clustering on a sparcc network matrix generated by filter_network(). and returns a tibble with clusters and taxon identifiers. Optionally, the network can be visualized.

Usage

cluster_network(network, min_n = 3, visualize = F)

Arguments

network

network network in matrix form.

min_n

minimum number of taxa per cluster, smaller clusters are filtered out.

visualize

whether to visualize the network clusters or not.