Filters the output of network() to a minimal threshold and transforms to matrix for downstream clustering or heatplot visualization.

Filters the output of network() to a minimal threshold and transforms to matrix for downstream clustering or heatplot visualization.

Usage

filter_network(network, threshold = 0.1, fdr = NULL, ...)

Arguments

network

network object generated by network().

threshold

absolute value of correlations below this threshold are

fdr

the threshold for false discovery rate (if pvalues are calculated for the network) filtered out.

...

Arguments passed on to stats::p.adjust

p

numeric vector of p-values (possibly with NAs). Any other R object is coerced by as.numeric.

method

correction method, a character string. Can be abbreviated.

n

number of comparisons, must be at least length(p); only set this (to non-default) when you know what you are doing!