Perform network inference with SparCC on a tidytacos object, after dropping rare taxa. See sparcc.

Perform network inference with SparCC on a tidytacos object, after dropping rare taxa. See sparcc.

Usage

network(
  ta,
  calculate_p = TRUE,
  min_occurrence = 0.05,
  taxon_name = taxon,
  sample_name = sample,
  ...
)

Arguments

ta

a tidytacos object

calculate_p

whether to calculate p-values or not. This can be time consuming. Iterations can be set with the R parameter and multiple cores through ncpus.

min_occurrence

Percentage of samples the taxon needs to be present in for it to be considered in the analysis.

iter

Number of iterations in the outer loop

inner_iter

Number of iterations in the inner loop

th

absolute value of correlations below this threshold are considered zero by the inner SparCC loop.